migrate id from Long to Integer

+++ src/dao/impl/AdminRepositoryImpls.java
@@ -3,7 +3,7 @@ package dao.impl;
 import dao.repository.AdminRepository;
 
 public class AdminRepositoryImpls extends UserRepositoryImpl implements AdminRepository {
-    public AdminRepositoryImpls(IdGenerator<Long> idGenerator) {
+    public AdminRepositoryImpls(IdGenerator<Integer> idGenerator) {
         super(idGenerator);
     }
