+++ src/dao/impl/RegisteredUseRepositoryImpl.java
@@ -3,7 +3,7 @@ package dao.impl;
 import dao.repository.RegisteredUserRepository;
 
 public class RegisteredUseRepositoryImpl extends UserRepositoryImpl implements RegisteredUserRepository {
-    public RegisteredUseRepositoryImpl(IdGenerator<Long> idGenerator) {
+    public RegisteredUseRepositoryImpl(IdGenerator<Integer> idGenerator) {
         super(idGenerator);
     }
 }
