+++ src/dao/repository/TicketRepository.java
@@ -2,5 +2,5 @@ package dao.repository;
 
 import model.entity.Ticket;
 
-public interface TicketRepository extends Repository<Long, Ticket>{
+public interface TicketRepository extends Repository<Integer, Ticket>{
 }
